Move root paths to utility.root
